SI-10066 Fix crash in erroneous code with implicits, dynamic
The compiler support in the typechecker for `scala.Dynamic` is
very particular about the `Context` in which it is typechecked.
It looks at the `tree` in the enclosing context to find the expression
immediately enclosing the dynamic selection. See the logic in
`dyna::mkInvoke` for the details.

This commit substitutes the result of `resetAttrs` into the tree
of the typer context before continuing with typechecking.