SI-10069 Fix code gen errors with array updates, Nothing
Crashes in ASM or VerifyErrors used to occur when assigning
an expression of type Nothing to an element of a primitive array.

This commit adapts the RHS of the assignment to the element
type to correct this. `adapt` contains logic to insert an
`ATHROW` of the slot of type `Nothing$`, which makes everything
line up. The subsequent array stores become dead code and are
dropped later on in code gen, so the test case compiles to:

  public void foo0(double[]);
    Code:
       0: bipush        42
       2: istore_2
       3: aload_1
       4: iconst_0
       5: aload_0
       6: invokevirtual #30                 // Method throwExpected:()Lscala/runtime/Nothing$;
       9: athrow

I found a similar bug in the emission of primitive unboxing and
fixed that too.