SI-8779 Enable inlining of code within a REPL session
The REPL has a long running instance of Global which outputs
classfiles by default to a VirtualDirectory. The inliner did not find
any of these class files when compiling calls to methods defined in
previous runs (ie, previous lines of input.)

This commit:

  - Adds a hook to augment the classpath that the optimizer searches,
    and uses this in the REPL to add the output directory
  - Fixes the implementation of `findClassFile` in VirtualDirectory,
    which doesn't seem to have been used in anger before. I've factored out
    some common code into a new method on `AbstractFile`.
  - Fixes a similar problem getSubDir reported by Li Haoyi
  - Adds missing unit test coverage.

This also fixes a bug in REPL autocompletion for types defined
in packages >= 2 level deep (with the `:paste -raw` command).
I've added a test for this case.