Improve performance of REPL autocompletion
The code used to fuzzily match, e.g, `declasses` with `getDeclaredClasses`
was exploring fruitless parts of the search space. The enclosed test
case was hanging the REPL.

This commit improves this by performing a prefix match of the unconsumed input
against the current chunk of the candidate before exploring the `inits`.

Fixes scala/scala-dev#271