2.12 -> 2.13 in the most important places
there are a lot of other places that could be updated, especially in
comments/docs, but the sooner we update them all, the more annoying
merge conflicts we'll have to suffer through. so at the moment let's
just hit the main ones

for now, just override it to use 2.12 modules+ScalaCheck. eventually
something will break but for now we'll get by