Consider a val/var to match a nullary method with the same name
The current definition of “matching” in the spec does not take the case
into account where a method matches a non-method member. I believe this
is necessary to make 5.1.4 (Overriding) correctly explain the case of
a val or var implementing an abstract method or overriding a concrete
method, both of which are allowed.