Test cases for SI-9111
The logic to resolve static nested classes changed 4e822d7, so the
description in the bug report (wrong imports) is no longer accurate.
The first example in the ticket is fixed, but not the second.

This commit adds tests for the status quo.

Removes a method that's no longer used since 4e822d7.