scala-xml dependency should be provided
Nothing should transitively depend on the scala-xml module except for
scaladoc and library-all.

Compiler needs scala-xml as a build dependency, but not at runtime.
The user is supposed to provide scala-xml as a dependency at runtime.
This fixes an issue where the classpath in SBT shadows the scala-xml
module, see scala/scala-module-dependency-sample#14.  The only
consequence of dropping the dependency at runtime is that scala-xml
always needs to be binary compatible.  That's the point of MiMa and
dbuild, no?