SI-9111 / SI-10107 resolution of static references in Java code
Scala puts static members of Java classes into the companion
object symbol. When parsing Java code, references to static
members need to be looked up in these companions.

In 2.11.x, a synthetic `import A._` was added to Java classes
having static members, which lead to some bugs (SI-9111). The
scheme was changed in 4e822d74be9 (2.12.x only): when a lookup
of `T` / `A.T` fails, the typer searches in the companion of the
current class.

This fixes some issues, but not all
  - In Java, `A.T` can refer to a static member defined in a
    parent of `A`
  - If a static `A.T` exists, but a parent of `A` defines a
    non-static member class `T`, the non-static one would be
    chosen (wrongly)

This commit pushes the logic to consider companion objects into
`FindMembers` and makes sure that for ever class in the base type
sequence, the companion is considered before moving to the next
base type.