SI-4788 clarify test cases
SI-4788 implemented support for `@Retention` annotations of Java
annotation classes: if a Java annotation class is itself annotated
`@Retention(value=SOURCE)`, it is not emitted to the annotated (Scala)
classfile.

However, the Java parser currently skips annotations altogether, so
if a Java annotation is passed to the Scala compiler in mixed compilation,
its `@Retention` not respected. Instead, uses of the annotation class
will be written to the classfile.

The existing tests for t4788 basically both tested separate compilation,
because partest's "mixed" compilation used to compile Scala sources twice
(the second time with the Java classfiles on the classpath). This was
fixed recently, uncovering the difference.

Additionally, the actual `t4788-separate-compilation` did not work as
intended because the filename groups were wrong (the `.scala` files should
all have been in the `_2` group.).