SI-10016 preserve annotation on existential
We'll just have to live with manifests not being
able to resolve the difference between `scala.List[_]` and `scala.List[Any]`
(dealiasing causes rewrapping causes existential extrapolation).
They are equivalent types after all (and manifests are deprecated).

A type tag will be more precise.

This commit reverts #1900, which caused this regression.
Recently, #5263 improved the situation a bit, but showed that
the original tweak was not a good idea (too many exceptions).
Also, I doubt using `=:=` will improve performance, even if it
reduces allocation a bit, as it's far more expensive than `eq`...