Do not instantiate TrieMap root updater for each instance
AtomicReferenceFieldUpdater is costly to instantiate, involving
call stack introspection, reflection and privileged operations.

It is recommended to pay this price only once, and then use an
updater for all objects.

Let's create a field in TrieMap companion object to hold an updater
and use that in the constructor. Unfortunately the companion object
is barred from instanting the updater by itself, as the updater
references a TrieMap private field. To overcome this, we define
a method by which the companion can request an updater from an
otherwise-defunct TrieMap instance.

This brings down the time to instantiate an empty TrieMap in local
testing from ~580ns to ~60ns, i.e. an improvement of one order
of magnitude.

Signed-off-by: Robert Varga <nite@hq.sk>