SI-1459 two bridges for impl of java generic vararg method
A Scala method that implements a generic, Java-defined
varargs method, needs two bridges:
  - to convert the collections for the repeated parameters (VBRIDGE)
  - to bridge the generics gap (BRIDGE)

Refchecks emits the varargs "bridges", and erasure takes care
of the other gap. Because a VBRIDGE was also an ARTIFACT,
it was wrongly considered inert with respect to erasure,
because `OverridingPairs` by default excluded artifacts.

Removed the artifact flag from those VBRIDGES, so that they
qualify for a real bridge. It would also work to include
VBRIDGE methods that are artifacts in BridgesCursor.