SI-9464 Clarify spec on no final trait
Without being too finicky with syntax,
say that `final abstract class` is OK
(even as a nested class, where it would
be an incomplete member) but not `final trait`.

Such a class might be a standard lib primitive,
or might be an implicit value that is only a
ludicrous marker and, if present, need only
have the null value.