scala-xml dependency should be provided
Nothing should transitively depend on the scala-xml module except for
scaladoc and library-all.

Compiler needs scala-xml as a build dependency, but not at runtime.

The user is supposed to provide scala-xml as a dependency at runtime.
This fixes an issue where the classpath in SBT shadows the scala-xml
module, see scala/scala-module-dependency-sample#14.

Unfortunately, scaladoc needs to be pulled out of the compiler.  This
is a first attempt at mangling the build so that scaladoc is a
separate jar.  Following on, SBT will need to know how to run scaladoc
as a separate jar.