Add checks against NoType
They are not relevant and their effect on performance is minimal, but
it's good to have them for consistency. We do check for `NoSymbol` as
well as `null` when it comes to symbols. It feels weird if we don't for
types.

Rewritten from sbt/zinc@fad19ce687a3071c1eb2a4bddc86492a58e04f87