Move dealiasing from Uncurry to Erasure
The motivation is to make the uncurry info tranform an identity
for more method signatures.

A few spots in erasure needed to be updated to account for
the possibility of aliases, by callng normalize and/or using
baseType rather than just assuming that the type args and class
type params line up.

Fixes scala/scala-dev#297