Use `ignoredSymbol` in API & reorder `isTopLevel`
`API` guards against null and `NoSymbol`s. This commit reuses those
checks in `GlobalHelpers` and rewrites `isTopLevel` so that it's easier
to see what the performed checks are.

This rewrite also reorders the checks for performance reasons. The way
`TopLevelTraverser` works is implicitly checking for `isNestedClass`
because it only traverses the trees that are inside package whose
symbols should have that package as an owner. `isTopLevel` is checked
this way, so it's a redundant check. However, it's not safe to remove it
from here because the semantics of `TopLevelTraverser` could change and
people would expect the implementation of `isTopLevel` to be stable.
Therefore, I've moved that check at the end of the check chain so that
only valid top level symbols pay the price of that check and we spare a
check in those symbols that are more likely to fail checks before.

Aside from this, `isStatic` is now checked before `isImplClass` because
non static symbols are more frequent and will therefore fail the check
before. Aside from this, `isImplClass` will always return `false` in
2.12 because the encoding of traits has changed.

Rewritten from sbt/zinc@5fbd7491d7fe3a4775710f1e8691c97f3ba74af8