SD-256 enable colored output by default on unix
`scala.color` now has 3 states: `true`, `false` and `auto`
(the default). `auto` allows colors if not on windows and if the shell
is interactive (as in, both stdin and stdout are a tty).

The autodetect works as expected when run via SBT too, and it can always
be overriden on the CLI or via JAVA_OPTS.