run partest from sbt always, command line never
these scripts are (I assume) unused these days and there's no reason
to maintain them. and it's risky to have two different ways of running
the same thing which could get out of sync with each other

sbt forks a JVM to run partest, so it's not like we need these scripts
in order to get a more isolated environment in that respect.