Add first version of JMH benchmarks
This commit introduces JMH benchmarks as requested by https://github.com/sbt/zinc/pull/225.

The strategy for the JMH benchmarks is the following:

1. Clone repository to be compiled and checkout at concrete commit (for
   reproducibility).
2. Generate sbt task for every subproject we want to test. This task
   will tell us information about the compiler options: sources and
   classpath (scalac options are integrated in next commit).
3. Execute the sbt tasks and get the output.
4. Instantiate a compiler fully integrated with the compiler bridge that
   has all the parameters given by the sbt output.
5. Run and enjoy.

The JMH benchmarks run this for every iteration, for now. However, they
only measure the running time of the **whole** compiler pipeline. There
is no reliable way to do it only for our Zinc phases since they are
tightly coupled to the compiler, so this is the best we can do.

The JMH benchmarks are only 2.12 compatible. This commit introduces a
benchmark for Shapeless and shows how easy it is to extend for other
codebases. It also adds some tests to make sure that the required
scaffolding works.

For that, we have to modify the accessibility of some methods in our
infrastructure, particularly `ScalaCompilerForUnitTesting`.

Rewritten from sbt/zinc@4a09b04ecf0c88332ba78c03b5ef3a59482c097e