SI-10187 Support mutation of mutable.HashMap in getOrElseUpdate
Scala 2.12.1 included optimizations to `HashMape.getOrElseUpdate`
to avoid recomputing the index in the hash table when adding an
the element.

However, this index could be stale if the callback added elements
to the map and triggered a resize.

This commit checks that the table is unchanged before reusing
the index, restoring the 2.12.0 behaviour.