Optimize `DependencySource` and `EnclosingNonLocal`
Having `DependencySource` and `EnclosingNonLocalClass` as wrapper in
both `ExtractUsedNames` and `Dependency` is nice, but it has a runtime
overhead.

The issue arises from the fact that caching is only performed at the
`currentOwner` level, and hence when we traverse the tree and change its
value, the cached instance is recomputed because the `currentOwner` does
not equal the owner of the previous cached instance. However, both
cached instances may resolve to the same non local class and recomputing
it is a waste of time.

Let's say we have this code:

```scala
object A {
  lazy val x = { ... }
  val y = { ... }
  def z(...) = { ... }
}
```

The traverser changes `currentOwner` when it enters `x`, `y` and `z` and
for every action we pay the initialization cost of another
`DependencySource` | `EnclosingNonLocalClass` instance.

This commit checks that the resolved non local class is the same, which
spare us several writes to disk. It also removes the runtime overhead of
`DependencySource` and `EnclosingNonLocalClass` by inlining its fields
as members of the enclosing `Traverser` class.

Rewritten from sbt/zinc@2f1002f7767e78c8b39d46d5c5f826313429f0c8