Improve test for "names in type tree"
The previous test was flaky because there was no guarantee that the
prefix of `BB` was correctly catched by the type traverser (the name of
`a` could have been recognised in the type of `foo` and not in `bar`).

The following PR makes sure that the prefix `c` is also captured.

Rewritten from sbt/zinc@83bcf506268bbcb2862906c4f04d63557a1fc1fe