Split up cache for ClassBTypes in two, according to origin
Split up the `classBTypeFromInternalName` map in two, depending if the
type was created from a symbol or classfile. This allows adding an additional
assertion: when getting a `ClassBType` for a symbol that's being compiled in
the current run, we want to assert that an existing, cached type was created
from that symbol, and not from a (matching, but old) classfile on the classpath.

Note that creating `ClassBTypes` from symbols is the ordinary case. In particular,
bytecode generation creats these types for every class being emitted. Types created
from classfiles are only used for type references not present in the soruce but
introduced by the inliner. At this later stage, all types for classes being compiled
are already created (and cached).