Remove some indirections in Global and Typer
This reduces the number of Java stack frames between the bottom
of the stack and the start of typer, and, more importantly,
at each level of recursion in the AST or in symbol completion.

The intent is to make the stacks easier to visually scan in
profilers and other tools that display them. I'm not expecting
that performance will improve, the JVM probably does a decent
jobs and inlining these chunks of the stack.

(cherry picked from commit 3d69134a50cc1e010e135377c54a7375db13ddb9)